separated type definitions
