+++ src/modules/claim/services/utils/Helper.ts
@@ -15,7 +15,7 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { objectAny, atomicObject } from "../../../../types";
+import { objectAny, atomicObject } from "../../../internal-types";
 
 
 export default class Helper {
