+++ src/modules/internal-types.ts
@@ -1,0 +1,26 @@
+/*
+ *  Copyright (C) 2020  Attila Aldemir <a_aldemir@hotmail.de>
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as published
+ *  by the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+import { IClaim } from "./../types";
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface IClaimService extends IClaim {}
+
+
+/*eslint-disable-next-line @typescript-eslint/no-explicit-any*/
+export type objectAny = {[name: string]: any};
+export type atomicObject = {[name: string]: string};
